base/instruction: fixed a number of bugs when dealing with structure operands.
Some issues were found related to locating a structure member that
is undefined, or bad calculations when trying to determine the offset
to an undefined structure member were fixed. Now we identify the
nearest member when one isn't found and use it to collect the correct
offset. Offsets and deltas that were less than zero were mistakenly
thrown out but are now fixed by ensuring they're included.